Major refactor of how the bar and player relate to one another
Big changes to the interaction of the bar and player modules.

The weapon speed update now uses the blocking flag to ensure that when the speed changes on an event that the old speed survives to be handled by the swing timer code if another event also triggers before the update.

Some of the bar element calculations now also happen outside code run on update, and are only recalced/shown/hidden as much as they need to be.