fixed infinite speed bug, KJ breath haste accounted for, crusader snapshotting redone
KJ breath buff is now detected and applied multiplicatively to the GCD calculation.

The attack speed in the API endpoint being returned as zero is now checked against and remedied.

Crusader snapshotting no longer freezes the attack speed entirely, but now instead applies a correction (which still breaks in the event crusader is activated and falls off in the same swing)