numerical tolerance function added, stupid edgecase in sotc application/falloff in same swing dealt with
Wrote a small function to compare two floats to one another within a small tolerance.

Dealt with an utterly obtuse edge case with SotC being applied then falling off in the same swing. The delay in the API updating means the correction to SotC attack speeds can be spikey.

bar is now properly click-throughable when locked