enabled a deadzone feature with corresponding options
A "deadzone" can now be enabled, which shows a partially transparent shaded region at the end of the bar where the player cannot change seals before they twist due to latency.

Added configurables to control the behaviour of this feature.

Also fixed an issue with the right text positioning being wrongly altered on bar movement