Multiple fixes for LibGlobalCooldown
Fix to ensure that Frostbolt rank 1 is used instead of Flash of Light if the player is a Paladin, to avoid Art of War procs reducing the polled cast time to zero.

Also fixed a bad not comparison that was preventing the projected GCD event from firing most of the time.