Implemented fix for range checkers not being loaded
On the first run of the game (only first time the game is started, not from logouts etc), LibRangeCheck fires two CHECKERS_CHANGED events. At the first, the lib is not yet ready. This caused a bug in the range finding loop on first load, that is not fixed by checking the checker funcs aren't returning nil on a loop before going to normal functionality