Fix to a zero division error that would happen to warriors with a shield equipped on load, attack speed bar element updates now trigger on weapon change
Shield being equipped on an initial load was attempting to set bar elements with an offhand speed of zero, preventing proper initialisation of the offhannd bar.

This is now fixed by setting a small but finite offhand speed if the offhand speed is zero.

In addition, the bar elements are now recalculated when the player's weapons change to ensure elements are updated promptly (especially deadzone for offhand in conjunction with the above change).